get run_capability_evals.sh up and running again
dataset.convert_to_questions takes a dataset as input. This is a bit
inelegant since I ended up passing the dataset to itself, so it could be refactored
in the future.

Running model_dump on the objects in the dataset causes an error since they
are already dicts, so I removed that.